Undo previous change, matchenv does NOT work on windows.
The reason is that matchenv works by looking at env variables, but in
4.02.3 (the version of ocaml bsb-native uses) there is a bug[1] with
Unix.execvpe which makes it not propagate the env variables.

That is one of the reasons we started using -bs-D to define variables
that bsb's ocaml could handle. It forces people to use ML for those
files but for now that's fine.

[1] https://caml.inria.fr/mantis/print_bug_page.php?bug_id=6903